Divide controller into internal and external.
