+++ server/controller/external.js
@@ -1,0 +1,8 @@
+/**
+    APIs for external servers.
+*/
+
+// sample.
+module.exports.sample = (req, res) => {
+    res.send('hello from external sample.')
+}
